requirements.txt
fastapi
uvicorn[standard]
streamlit
langchain-core
langchain-google-genai
langchain-neo4j
langchain-experimental
langchain-text-splitters
neo4j
pydantic
python-dotenv
crawl4ai
beautifulsoup4
tavily-python
